updates from code review
