+++ server/api/plaid.js
@@ -113,7 +113,6 @@ router.post('/transactions/:userId', async (req, res, next) => {
               )
               if (i === tokens.length - 1) {
                 res.json(transactions)
-                res.end()
               }
             }
           }
